window.c: Don't ungrab input after displaying the menu
We used to ungrab first to stop any existing grab and then grab after
showing the menu. That was broken in c680e904897dfb19331f8f296ab, which
moved the ungrab down below the grab, and as a result menus are now
shown without a grab.  This commit moves the grab back up.